CTCTRADERS-2490: Removed WhatDoYouWantToDo page and content. Refactored the Index page to become WhatDoYouWantToDo page.
