--- app/models/WhatDoYouWantToDoOptions.scala
@@ -1,47 +1,0 @@
-/*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package models
-
-import play.api.data.Form
-import uk.gov.hmrc.viewmodels.{MessageInterpolators, Radios}
-
-sealed trait WhatDoYouWantToDoOptions
-
-object WhatDoYouWantToDoOptions extends Enumerable.Implicits {
-
-  case object GBMovements extends WithName("gbMovements") with WhatDoYouWantToDoOptions
-  case object NorthernIrelandMovements extends WithName("northernIrelandMovements") with WhatDoYouWantToDoOptions
-
-  val values: Seq[WhatDoYouWantToDoOptions] = Seq(
-    GBMovements,
-    NorthernIrelandMovements
-  )
-
-  def radios(form: Form[_]): Seq[Radios.Item] = {
-    val field = form("value")
-
-    val radioButtons = Seq(
-      Radios.Radio(msg"whatDoYouWantToDo.gbMovements", GBMovements.toString),
-      Radios.Radio(msg"whatDoYouWantToDo.northernIrelandMovementsText", NorthernIrelandMovements.toString)
-    )
-
-    Radios(field, radioButtons)
-  }
-
-  implicit val enumerable: Enumerable[WhatDoYouWantToDoOptions] =
-    Enumerable(values.map(v => v.toString -> v): _*)
-}
