+++ app/navigation/Navigator.scala
@@ -29,13 +29,13 @@ class Navigator @Inject()() {
   private val normalRoutes: Page => UserAnswers => Call = {
     case _ =>
       _ =>
-        routes.IndexController.onPageLoad()
+        routes.WhatDoYouWantToDoController.onPageLoad()
   }
 
   private val checkRouteMap: Page => UserAnswers => Call = {
     case _ =>
       _ =>
-        routes.IndexController.onPageLoad()
+        routes.WhatDoYouWantToDoController.onPageLoad()
   }
 
   def nextPage(page: Page, mode: Mode, userAnswers: UserAnswers): Call =
