+++ conf/app.routes
@@ -4,10 +4,7 @@ GET        /assets/*file                                                      co
 
 GET        /                                                                  controllers.RedirectController.onPageLoad()
 
-GET        /what-do-you-want-to-do                                            controllers.WhatDoYouWantToDoController.onPageLoad()
-POST       /what-do-you-want-to-do                                            controllers.WhatDoYouWantToDoController.onSubmit
-
-GET        /index                                                             controllers.IndexController.onPageLoad
+GET        /what-do-you-want-to-do                                            controllers.WhatDoYouWantToDoController.onPageLoad
 
 GET        /this-service-has-been-reset                                       controllers.SessionExpiredController.onPageLoad
