+++ conf/application.conf
@@ -108,7 +108,7 @@ microservice {
 
       features {
         departureJourney     = true
-        isNIJourneyEnabled   = false
+        isNIJourneyEnabled   = true
       }
 
       tracking-consent-frontend {
