+++ conf/views/includes/layout.njk
@@ -24,7 +24,7 @@
   {{ hmrcHeader({
     homepageUrl: "https://www.gov.uk",
     serviceName: messages("site.service_name"),
-    serviceUrl: "/manage-transit-movements/index",
+    serviceUrl: "/manage-transit-movements/what-do-you-want-to-do",
     containerClasses: "govuk-width-container",
     assetsPath: "/assets",
     phaseBanner: {
