--- conf/views/index.njk
@@ -1,64 +1,0 @@
-{% extends "includes/layout.njk" %}
-
-{% from "macros/title.njk"                         import title %}
-
-{% block pageTitle %}
-{{ (messages("index.title")) }}{% endblock %}
-{% block mainContent %}
-
-  <div class="govuk-main-wrapper">
-    <div class="govuk-grid-row">
-      <div class="govuk-grid-column-two-thirds">
-
-        <h1 class="govuk-heading-xl">
-          {{ messages("index.heading") }}
-        </h1>
-
-        <h2 class="govuk-heading-m">
-            {{ messages("index.subHeading.arrival") }}
-        </h2>
-        <p class="govuk-body">
-            <a id= "make-arrival-notification" href={{declareArrivalNotificationUrl}}>
-                {{ messages("index.link.arrival.makeArrival") }}
-            </a>
-        </p>
-        <p class="govuk-body">
-        {% if hasArrivals %}
-            <a id= "view-arrival-notifications" href={{ viewArrivalNotificationUrl }}>
-                {{ messages("index.link.arrival.viewArrival") }}
-            </a>
-        {% else %}
-            {% if arrivalsAvailable %}
-                {{ messages("index.arrival.noArrivals") }}
-            {% else %}
-                {{ messages("index.arrival.arrivalsUnavailable") }}
-            {% endif %}
-        {% endif %}
-        </p>
-         
-        <h2 class="govuk-heading-m">
-            {{ messages("index.subHeading.departure") }}
-        </h2>
-        <p class="govuk-body">
-            <a id= "make-departure-declaration" href={{ declareDepartureDeclarationUrl }}>
-                {{ messages("index.link.departure.makeDeparture") }}
-            </a>
-        </p>
-        <p class="govuk-body">
-        {% if hasDepartures %}
-            <a id= "view-departure-declarations" href="{{ viewDepartureNotificationUrl }}">
-                {{ messages("index.link.departure.viewDepartures") }}
-            </a>
-        {% else %}
-            {% if departuresAvailable %}
-                {{ messages("index.departure.noDepartures") }}
-            {% else %}
-                {{ messages("index.departure.departuresUnavailable") }}
-            {% endif %}
-        {% endif %}
-        </p>
-      </div>
-    </div>
-  </div>
-
-{% endblock %}
